refactor: successful-bid-topic 토픽 consume 로직 캐스팅 수정(#44)

+++ src/main/java/com/skyhorsemanpower/payment/kafka/KafkaConsumerCluster.java
@@ -4,9 +4,11 @@ import com.skyhorsemanpower.payment.kafka.dto.AlarmDto;
 import com.skyhorsemanpower.payment.payment.application.PaymentService;
 import com.skyhorsemanpower.payment.payment.dto.PaymentListResponseDto;
 import com.skyhorsemanpower.payment.payment.vo.PaymentReadyVo;
+
 import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 import java.util.List;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -25,16 +27,17 @@ public class KafkaConsumerCluster {
 
     @KafkaListener(topics = "successful-bid-topic", groupId = "${spring.kafka.consumer.group-id}")
     public void consumeBidder(@Payload LinkedHashMap<String, Object> message,
-        @Headers MessageHeaders messageHeaders) {
+                              @Headers MessageHeaders messageHeaders) {
+
         log.info("consumer: success >>> message: {}, headers: {}", message.toString(),
-            messageHeaders);
+                messageHeaders);
 
         //message를 PaymentReadyVo로 변환
         PaymentReadyVo paymentReadyVo = PaymentReadyVo.builder()
-            .auctionUuid(message.get("auctionUuid").toString())
-            .memberUuids((List<String>) message.get("memberUuids"))
-            .price(BigDecimal.valueOf((Integer) message.get("price")))
-            .build();
+                .auctionUuid(message.get("auctionUuid").toString())
+                .memberUuids((List<String>) message.get("memberUuids"))
+                .price(BigDecimal.valueOf((Double) message.get("price")))
+                .build();
 
         log.info("consumer: success >>> paymentReadyVo: {}", paymentReadyVo.toString());
 
@@ -43,16 +46,16 @@ public class KafkaConsumerCluster {
 
     @KafkaListener(topics = "event-preview-topic", groupId = "${spring.kafka.consumer.group-id}")
     public void consumeEventPreview(@Payload LinkedHashMap<String, Object> message,
-        @Headers MessageHeaders messageHeaders) {
+                                    @Headers MessageHeaders messageHeaders) {
         String auctionUuid = message.get("auctionUuid").toString();
         List<PaymentListResponseDto> paymentListResponseDtos = paymentService.findCompletePayments(
-            auctionUuid);
+                auctionUuid);
         List<String> memberUuids = paymentListResponseDtos.stream()
-            .map(PaymentListResponseDto::getMemberUuid).toList();
+                .map(PaymentListResponseDto::getMemberUuid).toList();
 
         producer.sendMessage(Topics.NOTIFICATION_SERVICE.getTopic(), AlarmDto.builder()
-            .receiverUuids(memberUuids)
-            .eventType("payment")
-            .message("행사 시작까지 24시간 남았어요.").build());
+                .receiverUuids(memberUuids)
+                .eventType("payment")
+                .message("행사 시작까지 24시간 남았어요.").build());
     }
 }
